requirements.txt
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
category_encoders==2.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
Choco==1.0.5
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
executing==2.0.1
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.1
fonttools==4.52.1
fqdn==1.5.1
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.5
lark==1.3.0
lightgbm==4.6.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.1.3
mlxtend==0.23.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
packaging==24.0
pandas==2.3.0
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
pipenv==2024.4.0
platformdirs==4.2.1
plumbum==1.9.0
ply==3.11
polars==1.33.0
prometheus_client==0.23.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pygame-ce==2.5.5
PyGithub==2.6.1
Pygments==2.17.2
PyJWT==2.10.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
pytz==2024.1
pywin32==306
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==26.0.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.6.0
six==1.16.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20251008
typing_extensions==4.13.1
tzdata==2024.1
ucimlrepo==0.0.7
umap==0.1.1
umap-learn==0.5.9.post2
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.28.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xgboost==3.0.1
